klte-common: import correct libCB.so from the Adreno package we are using.
* During some housekeeping I did a diff of these libs with the archive
I had saved that I downloaded when I updated these. Somehow I swapped
libCB.so. The one that we were using was incorrect and not part of this
package. It was from the stock TW KK ROM and from the KK Adreno branch

* This corrects that derp

* Adreno package version:
AU_LINUX_ANDROID_LNX.LA.3.5.2.2_RB1.04.04.04.087.028_msm8974

Change-Id: I2f3266a42483fc1692c6ef897a17b8ad74d08814